+++ src/utils/time.h
@@ -1,0 +1,11 @@
+/**
+* @copyright 2026 - Max Bebök
+* @license MIT
+*/
+#pragma once
+#include <string>
+
+namespace Utils::Time
+{
+  std::string getDateTimeStr();
+}